backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose
bcrypt
passlib
